Do not sign-extend 0xFF to EOF when reading from buffer
For background, see: https://github.com/skeeto/pdjson/issues/37

+++ libpdjson5/pdjson5.c
@@ -223,9 +223,11 @@ source_peek (pdjson_stream *json)
 {
   struct pdjson_source *source = &json->source;
 
+  // Make sure we don't sign-extend 0xFF to EOF.
+  //
   return source->tag == PDJSON_SOURCE_BUFFER
     ? (source->position != source->source.buffer.length
-       ? source->source.buffer.buffer[source->position]
+       ? (int)(uint8_t)source->source.buffer.buffer[source->position]
        : EOF)
     : source_peek_slow (json, source);
 }
@@ -274,9 +276,11 @@ source_get (pdjson_stream *json)
 {
   struct pdjson_source *source = &json->source;
 
+  // Make sure we don't sign-extend 0xFF to EOF.
+  //
   return source->tag == PDJSON_SOURCE_BUFFER
     ? (source->position != source->source.buffer.length
-       ? source->source.buffer.buffer[source->position++]
+       ? (int)(uint8_t)source->source.buffer.buffer[source->position++]
        : EOF)
     : source_get_slow (json, source);
 }
